requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.5
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autopep8 @ file:///home/conda/feedstock_root/build_artifacts/autopep8_1677845287175/work
autosub @ git+https://github.com/agermanidis/autosub.git@d32389cb76e63ec6959111c3f989a72f36f726fe
bing-image-downloader==1.1.2
boltons @ file:///croot/boltons_1677628695607/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175566073/work
brotlipy==0.7.0
cachetools==5.3.0
certifi==2023.5.7
cffi @ file:///croot/cffi_1670423216006/work
chardet==5.1.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
conda==23.3.1
conda-content-trust @ file:///tmp/abs_d9e8d879-c7bd-425e-87af-59f595eb114bfmj_f2_x/croots/recipe/conda-content-trust_1658126375626/work
conda-package-handling @ file:///croot/conda-package-handling_1672865005623/work
conda_package_streaming @ file:///croot/conda-package-streaming_1670508156500/work
cryptography @ file:///croot/cryptography_1677533078807/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dill==0.3.6
et-xmlfile @ file:///home/conda/feedstock_root/build_artifacts/et_xmlfile_1674664118162/work
exceptiongroup==1.1.1
ffmpeg-python==0.2.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
google-api-core==2.11.0
google-api-python-client==2.87.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-images-download==2.8.0
googleapis-common-protos==1.59.0
gTTS @ file:///home/conda/feedstock_root/build_artifacts/gtts_1673886701848/work
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.14.1
idna @ file:///croot/idna_1666125574438/work
imageio @ file:///home/conda/feedstock_root/build_artifacts/imageio_1683031833737/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1673483481485/work
isort==5.12.0
Jinja2==3.1.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
lazy-object-proxy==1.9.0
llvmlite==0.40.0
MarkupSafe==2.1.2
mccabe==0.7.0
more-itertools==9.1.0
moviepy @ file:///home/conda/feedstock_root/build_artifacts/moviepy_1665160419595/work
mpmath==1.3.0
multidict==6.0.4
mutagen @ file:///home/conda/feedstock_root/build_artifacts/mutagen_1665355730162/work
networkx==3.1
numba==0.57.0
numexpr @ file:///croot/numexpr_1683221807349/work
numpy @ file:///croot/numpy_and_numpy_base_1682520579168/work
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.7
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
openpyxl @ file:///home/conda/feedstock_root/build_artifacts/openpyxl_1632332968451/work
outcome==1.2.0
packaging @ file:///croot/packaging_1678965316161/work
pandas==1.5.3
pandas-stubs @ file:///home/conda/feedstock_root/build_artifacts/pandas-stubs_1682984395684/work
Pillow==9.4.0
platformdirs==3.5.1
pluggy @ file:///opt/conda/conda-bld/pluggy_1648109264631/work
proglog==0.1.9
progressbar==2.5
progressbar2==4.2.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle @ file:///home/conda/feedstock_root/build_artifacts/pycodestyle_1669306857274/work
pycosat @ file:///croot/pycosat_1666805506522/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydub @ file:///home/conda/feedstock_root/build_artifacts/pydub_1615612442567/work
pylint==2.17.4
pyOpenSSL @ file:///croot/pyopenssl_1677607690928/work
pyparsing==3.0.9
PySocks @ file:///home/ktietz/ci_310/pysocks_1643727729738/work
pysrt==1.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
python-utils==3.5.2
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
PyYAML==6.0
regex==2023.5.5
requests @ file:///croot/requests_1678709725998/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml @ file:///croot/ruamel.yaml_1666307131939/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1666302244777/work
selenium==4.9.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition @ file:///home/conda/feedstock_root/build_artifacts/speechrecognition_1678801698152/work
sympy==1.12
tiktoken==0.3.3
tokenizers==0.13.3
tomli @ file:///home/conda/feedstock_root/build_artifacts/tomli_1644342247877/work
tomlkit==0.11.8
toolz @ file:///croot/toolz_1667464081129/work
torch==2.0.1
tqdm @ file:///croot/tqdm_1679561868826/work
transformers==4.29.2
trio==0.22.0
trio-websocket==0.10.2
types-pytz @ file:///home/conda/feedstock_root/build_artifacts/types-pytz_1680176259362/work
typing_extensions==4.6.0
uritemplate==4.1.1
urllib3 @ file:///croot/urllib3_1680254685414/work
whisper-openai==1.0.0
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zstandard @ file:///croot/zstandard_1677013116773/work
